Extract join_event command into module
