+++ lib/food_bot/slack_bot.ex
@@ -1,8 +1,8 @@
 defmodule FoodBot.SlackBot do
   use Slack
 
-  alias FoodBot.{Repo, Event, Order}
-  import Ecto.Query
+  alias FoodBot.{Repo, Order}
+  import FoodBot.CommandHelpers
 
   def start_link() do
     Slack.Bot.start_link(__MODULE__, %{}, Application.get_env(:food_bot, :slack_bot_token))
@@ -29,55 +29,11 @@ defmodule FoodBot.SlackBot do
   end
   def handle_event(_, _, state), do: {:ok, state}
 
-  def latest_events_text do
-    Event
-    |> order_by(desc: :inserted_at)
-    |> limit(5)
-    |> Repo.all
-    |> Enum.map(&(" - #{&1.name}"))
-    |> Enum.join("\n")
-  end
 
   def handle_command(cmd, rest \\ nil, state \\ %{})
-  def handle_command("join_event", nil, state) do
-    {
-      """
-      Sorry, you didn't provide an event name. Is it one of these?
-      #{latest_events_text()}
-      """,
-      state
-    }
-  end
+
   def handle_command("join_event", name, state) do
-    event = Event
-            |> preload(:food_sources)
-            |> where(name: ^name)
-            |> limit(1)
-            |> Repo.one
-
-    case event do
-      nil ->
-        {
-          """
-          Sorry, I can't find event "#{name}". Is it one of these?
-          #{latest_events_text()}
-          """,
-          state
-        }
-
-      event ->
-        food_sources_text = event.food_sources
-                            |> Enum.map(&(" - #{&1.name}: #{&1.url}"))
-                            |> Enum.join("\n")
-
-        {
-          """
-          You joined event "#{name}". You can order from:
-          #{food_sources_text}
-          """,
-          Map.put(state, :event, event)
-        }
-    end
+    FoodBot.JoinEventCommand.execute(name, state)
   end
   def handle_command("order", nil, state) do
     { "Sorry, you didn't provide an order.", state }
